+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ##### Enhancements
 
 - Added `--write-results <file>` option.
-- Added `markdown` output format.
+- Added `github-markdown` output format.
 
 ##### Bug Fixes